CBD-86 Ensure queries always get consistent snapshots
After the state transition optimizations done in the commit
ad151b0509043cf7e9c462b4a1e24b18e86721d3, it was possible
that callers could get a view group snapshot that was not
fully consistent according to the callers' desired set of
indexed partitions. This is because a subset of the caller's
desired partitions could still be in the group's pending
state transition - when this happens, block the caller until
the pending state transition is applied or mutates in such a
way that it no longer contains any of the caller's desired
indexed partitions.

This prevents ocassional query "inconsistent_state" errors
that started happening more often after the commit mentioned
above.

Change-Id: I63423c95c3afa8d69419ce88f02d257fa92c44ec
Reviewed-on: http://review.couchbase.org/15359
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>