CBD-104 Remove unnecessary header fields
The update_seqs and purge_seqs fields, which are long lists
(1024 entries) with elements that have the form
{PartitionId::integer(), Seq::integer()}, and were part of the
view states, were not really used. We need one of this lists only,
which applies to all views in the view group and already exist as
top level fields of the index header record.

This reduces memory copying of data structures when a view group
is passed between processes (query requests, compactor, updater,
cleaner) and results in smaller headers.

Change-Id: I117b0e4ec225c1fd064aa8ab106eed1fc2eadcf9
Reviewed-on: http://review.couchbase.org/15617
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>