MB-5241 Add missing replica cleanup operation
Currently if a replica partition transitions to the active
state, a process that transfer the partition's data from the
replica index to main index is started.
If before that process finishes someone asks to transition the
partition to the passive state, the proper actions happen on the
main index, but nothing is done on the replica index, leaving the
partition's data in the replica index. This doesn't seem to be
harmful however, it only consumes unnecessary disk space and IO
(due to automatic replica index updates).

Change-Id: If513a61247ff888da431d34bd9ac5c29cc7b1293
Reviewed-on: http://review.couchbase.org/15665
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>