CBD-106 Lighter group snapshots for non-debug requests
Instead of sending a big fat group snapshot full of debug
information and unncessary fields for querying to any request,
use a much more lightweight version to send to non-debug mode
requests. This reduces term copying when sending the snapshots
to requesters and reduces contention on the main view group
process.

Change-Id: I09a5ac8f676911b61f5e6465ead45b1d1c92809f
Reviewed-on: http://review.couchbase.org/15735
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>