CBD-137 Fix active/passive bitmasks when compactor finishes
When the compactor finishes, we set his group snapshot's replicas on
transfer field to the current set of replicas marked as on transfer in
the current group, but we don't update the active and passive bitmasks.
This is wrong, because if both groups' replicas on transfer are different,
it means some partitions had a transition from the passive state to the
active state, therefore the passive and active bitmasks must be updated in
the group snapshot received from the compactor.

Change-Id: Icae3afc9636705fc63acfb8d4ea7e53d0815daf7
Reviewed-on: http://review.couchbase.org/15856
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>