CBD-104 Commit and fsync header in compactor process
After a group compaction finishes, the header was written and
the new file fsync'ed in the main view group process - this can
take a long period of time, which blocks the main view group
process, making view query requests wait for the write/fsync
to finish in order to get a group snapshot from the main view
group process and start folding a view.

The header is now written and the new file fsync'ed by the
compactor process. The main view group process will only write
a new header and fsync the new file if some important metadata
changed while the compaction was running - this happens for
the non-steady scenario, namely during (and after for a short
period) cluster rebalances and failovers. If this happens, the
main view group process is only writing and fsync'ing one header,
which is cheap - the compactor process alreadt fsync'ed all the
data and one obsolete header.

Change-Id: Ib004ea76b12ec1c3f3c891f13746f67b4f9b7d29
Reviewed-on: http://review.couchbase.org/15913
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>