CBD-104 Checkpoint indexing progress in the updater
When building/updating an index, do the checkpoints (header writes)
through the updater process, and only send a new group snapshot
when a checkpoint is performed (every 5 seconds). This reduces
contention in the main view group process and avoids the overhead
of copying group snapshot terms after every batch is written to
the index.

Change-Id: I4af0b5c7be076bc41313e1593969bb4b0d5a59d2
Reviewed-on: http://review.couchbase.org/15914
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>