MB-5270 Ignore deletes of databases in pending cleanup
If a database is in the list of pending cleanup partitions,
ignore db deleted events for this database (just like if it
was not pending).

If the updater starts and its input group snapshot has a
pending transition, perform a full cleanup before reading
changes from the partitions to update the index. This is
needed to ensure it doesn't open databases that were silently
recreated and starts indexing data from them without removing
first all previous data originated from those databases.
This full cleanup step is interruptible (just like everywhere
else) and the partition cleanup list consists of the union of
the current cleanup list and the pending transition cleanup list.

Change-Id: Ibc8bc29fdfa3865c0b81a5d8906c230c1a435727
Reviewed-on: http://review.couchbase.org/15929
Reviewed-by: Damien Katz <damien@couchbase.com>
Tested-by: buildbot <build@couchbase.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>