CBD-171 Asynchronous view merger queuing behaviour
After producing a row, when queuing it, a folder process was
blocked until the main merger process consumed the row.
This was non optimal, reducing parallelism. Instead of blocking
the producer, allow it to produce the next row until the previous
one in consumed by the main merger process.  This is particularly
beneficial when the folder is folding a remote view, where socket
IO and JSON decoding is time consuming.

Change-Id: Ifdf3237d1cee153b671a9908aef53db92d2be5e7
Reviewed-on: http://review.couchbase.org/16237
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>