CBD-105 Yet more efficient way to find if index needs update
This is a refinement over the previous CBD-105 change.
The database sequence numbers are now kept in an orddict
(sorted proplists) so when a caller requests for the current
database sequences, we no longer build an orddict per such
request.
This also simplifies the db set module, namely it removes the
categorization of active and passive partitions - this was
legacy code no longer needed.

Change-Id: I6230b6b42ac6c325b876e9e475152c0e1fc47033
Reviewed-on: http://review.couchbase.org/16326
Reviewed-by: Damien Katz <damien@couchbase.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>