CBD-335 Close log file if compactor restarts
If the compactor dies while the updater is running, the next
time the compactor starts, if the updater is still running,
the compactor will delete the log file, but the updater will
be writing to the old, no longer "openable" file for a while
and it will eventually crash when the compactor tells it to
start logging updates to the log file - this is due to an
assertion that asserts there's no current log file when the
updater receives a start logging request from the compactor.

The solution here is for the updater to close the old file
and create a new one with the same name when it receives a
message from the compactor telling it to start writing updates
to the log file. This avoids an updater crash which would also
make the compactor crash (next time both were restarted there
would be no problems however).

Change-Id: Ia003943cdd2988a719566e6a575c39d6e66bdb80
Reviewed-on: http://review.couchbase.org/17017
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>