MB-5511 Avoid deadlock between updater and view group
Depending on timing issues, it's possible that, at the same
time the view group is attempting to stop the updater, the
updater is doing a synchronous call to the view group to
tell it about the compaction log latest eof. This results in
a deadlock.
The notification of the log eof can be done asynchronously
wihtout any issues, which avoids the deadlock.

Change-Id: If1a003aba869b908965b972a3f3a8727395c8ddd
Reviewed-on: http://review.couchbase.org/17116
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Damien Katz <damien@couchbase.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>