CBD-279 Parallel shutdown of view folders and queue
Also, the usage of couch_util:shutdown_sync/1 now consumes
exit messages, which the merger needs to to analyze in order
to raise an error to the client when something goes wrong.
If the exit reason of any folder (or the queue) is 'shutdown',
don't treat as an error, after all, these processes were
explicitly shutdown.

Change-Id: I560e8c1c9443db474baf65d75781a1c442b6fac7
Reviewed-on: http://review.couchbase.org/17589
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>