CBD-69 Periodically update set view indexes
There's now an internal daemon that periodically (default period
is 5 seconds) triggers an update on all indexes if there are at
least N new changes to index (default N = 5000).
A periodicity of 0, or a minimum number of change of 0, disables
the automatic index updates, ensuring exactly the same behaviour
as before this change.

Change-Id: I15b49b6b27063319651908b12d6533797049198e
Reviewed-on: http://review.couchbase.org/17702
Reviewed-by: Damien Katz <damien@couchbase.com>
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>