CBD-74 Update total_changes in index task status as well
The total number of changes to index can change (increase) during
indexing - new documents, document updates and/or document deletes
are happening in parallel.
Therefore the initial number of changes missing, to apply to an index,
that the indexer calculates when it starts, may not be valid anymore
until it finishes.

Change-Id: If3b9ba7bd1089ef13040267097ab6e9cd9f68134
Reviewed-on: http://review.couchbase.org/17884
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>