MB-5398 Remove timeout when requesting next data chunk
This timeout existed to deal with misbehaviour of the http
library ibrowse.
Ibrowse is no longer used, lhttpc replaced it and doesn't
seem to suffer the same issue, as it isn't modeled with
a gen_server, passive sockets and complex state transitions
as ibrowse is (just uses plain old socket recv calls when
a data chunk is request by user).

This affected view queries with stale=false option in a multi-node
scenario where the client would get timeout errors for remote
nodes.

Lhttpc patch submitted and merged upstream:

https://github.com/esl/lhttpc/pull/18

Change-Id: I28a0a1607420c3df717dbdc2d3dfa6d2a0bb58bb
Reviewed-on: http://review.couchbase.org/17980
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>