CBD-438 More efficient btree copy
While folding the items (leaf values), avoid dict lookups and stores
for each item, reducing garbage creation and decreasing process
reductions. For upper levels, use an array instead of a dict, which
is more efficient than dicts (gets and sets are about twice as fast).

For a btree with 5M items, the total copy time was reduced from about
5 minutes to 3 minutes.

Change-Id: I73fd40709f2a10620548dc993be19df1bc955e8f
Reviewed-on: http://review.couchbase.org/18143
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>