MB-5865 Fix couch_server crash on database delete
If a database is deleted while it's being open, the process
that is opening the database is shutdown but it might have
already sent an open_result message to couch_server. When
couch_server receives that message, it will crash because it
expected the ets table to have an entry for that database, but
it does no longer have one because the delete request removed
that entry when it shutdown the opener process.

This lead to a crash of couch_server which got propagated to
several processes (processes opening/deleting other databases
and all previously open databases).

Change-Id: I44c08b5d77d299fc48e7d3f09a5d4d09af7080ae
Reviewed-on: http://review.couchbase.org/18207
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Reviewed-by: Damien Katz <damien@couchbase.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>