CBD-425: Use ale logger for MapReduce errors
If we run the full stack, the errors during the MapReduce phase
will be logged into a separate log. The ale logger (which is
used by ns_server) will be used, to make it possible to display
the errors on the UI.

If the ale logger isn't available, the errors will end up in
the normal couchdb log file. This is e.g. the case when the
couchdb tests are run.

The mapreduce ale logger is called directly without using the
syntax transformation. This is done so CouchDB doesn't have
a build time dependency on ale.

This commit needs to have the mapreduce logger enabled in
ns_server (http://review.couchbase.org/18292) in order to see
the errors logged with ale to a separate file.

Change-Id: I4ff7724f83da018631e4324da56ed226235fda2e
Reviewed-on: http://review.couchbase.org/18293
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Volker Mische <volker.mische@gmail.com>