CBD-386 / CBD-444 More reliable stale=false snapshot requests
When such group snapshot requests come, grab the sequence number
of all active partitions and reply to the caller immediately
if the those were already indexed, otherwise just store those
sequence numbers and unblock/reply to the caller once the index
is up to date regarding those sequence numbers.
Previously callers were unconditionally unblocked once the updater
finished, which was not semantically correct, specially if a request
came after the updater started (common in an heavy updates scenario
with multiple clients doing stale=false queries).

Callers will also be blocked if any of the active partitions are
paused for indexing, and unblocked once they are unpaused and the
currently indexed sequence numbers are greater or equal to the
sequence numbers the caller got when it made the group snapshot
request (CBD-386).

Change-Id: I2e852f8c6e87a6a9909908e2d931bab919e46085
Reviewed-on: http://review.couchbase.org/18411
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>