CBD-457 Faster updater stop on state transitions
Currently on state transitions, we must stop and restart the updater.
We gracefully shutdown the updater, waiting for it to finish processing
its current batch (5000 docs or max of 512Kb of KV pairs). This can
take a significant time if the index has many views with very deep
btrees each, making ns_server wait a long time.

This change just shutdowns the updater, with the tradeoff of losing
at most 5 seconds of updates, which will be processed again when the
updater is restarted. This has no side effects other than some
recomputations.

Change-Id: I207a9dfaeefd8553f64df0a33cea00926cbdfd18
Reviewed-on: http://review.couchbase.org/18538
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>