MB-6039 Don't crash if seq number is lower than indexed one
After recent query performance work (CBD-402), getting the current
seq numbers for the databases is done in a not fully synchronous
fashion for stale=update_after requests or when verifying if replica
index needs to be updated. This means that if the updater just finished,
it could have seen a higher seq number for a database than the one
we get after from the db set (because it's not done in a fully
synchronous way, except for stale=false requests).

Change-Id: Id9ebf86383b1ed7b5fb1553c762fd9249df6359f
Reviewed-on: http://review.couchbase.org/19007
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>