CBD-426 Update ejson collator
This is the latest version, which is thread safe (ICU collators are not
guaranteed to be thread safe nor reentrant) and has several other
improvements.

This was reviewed previously and merged, but somehow not present anymore
(maybe there was some git/gerrit issue that caused the rollback?).
Fortunately this last and better version was still intact in Apache CouchDB.

This is used for old CouchDB views (development views) and part of this
will be re-used by the new raw JSON collator (CBD-426).

Change-Id: Ibc65dfb6c71ae2af2af6ad64acf0790bc5d72e14
Reviewed-on: http://review.couchbase.org/19016
Reviewed-by: Damien Katz <damien@couchbase.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>