MB-6056 Fix distributed skip and limit
When used together, limit can not be propagated as it is without
summing the value of skip to it. We don't know in advance how
many rows are discarded from each node, so in worst case N (where
N is limit value) rows are discarded from a node, effectively
counting as a limit of 0 for that node.
Regression introduced in cf25a4b26c923645eb9acaee1b7b282f71af5c21.

Change-Id: Ia129cc3463d79e2b1e5f04ca8e5ec69320734db9
Reviewed-on: http://review.couchbase.org/19034
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>