MB-6208 Don't crash in before_database_delete call
A crash in the synchronous before_database_delete would lead to
the terminate callback be invoked which in turn can lead to
deadlocks, as it it synchronously shutdowns view groups.
Crash caused by the fact that when a view group dies, the ets
entries for its aliases were not deleted.

Change-Id: I7f04ff324711f17e948058bb54d2f727c29de0b6
Reviewed-on: http://review.couchbase.org/19658
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>