CBD-511 Get all ddoc aliases on view group restart
If a view group restarts, grab the id of all design documents
with the same signature (aliases). We were only populating the
ets tables when ddocs are created (or deleted), so in the case
of a view group restart (or server restart) we would no entries
for the aliases in the ets table.

Change-Id: Ibfa913ba49f9585c205e46cbc7f69032bb011e85
Reviewed-on: http://review.couchbase.org/19700
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>