MB-6415 Delete index files on view group terminate
A quick succession of ddoc updates, updating the ddoc back
to a previous version (while we're here in terminate), may lead
us to a case where it should start clean but it doesn't, because
couch_set_view:cleanup_index_files/1 was not invoked right before
the last reverting ddoc update, or it was invoked but did not finish
before the view group got spawned again. Problem here is during
rebalance, where one of the databases might have been deleted after
the last ddoc update and while or after we are in terminate callback,
so we must ensure we don't leave old index files around.

Change-Id: Ieb77ed59ac4de4a62882c711604bf360d6f46dfd
Reviewed-on: http://review.couchbase.org/20164
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>