MB-6541 Notify update listeners when view compaction finishes
Due to view compaction retries, it's possible the compacted
group has caught up with the updater's group and it's actually
more recent than the view group's current group snapshot.
In this case notify the update listeners, as a new index update
might not be triggered until new changes are applied to the
vbucket databases and a subsequent query with stale=false or
stale=update_after arrives.

Change-Id: I455f699bb5ca5c31ab30c365df960e115ab0d448
Reviewed-on: http://review.couchbase.org/20587
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>
Reviewed-by: Volker Mische <volker.mische@gmail.com>