MB-6480 Use file:advise/4 when finishing initial index build
When build the btrees at the end of the initial index build,
we're reading sequentially from a file which we will delete
when this step finishes. Reading this file will make the OS
page in all its blocks into the page cache, which can trash
the entire page cache if the file is too big, or we have
many indexes building or indexes with many views for large
datasets.

This change simply periodically tells the OS it can purge
pages from the cache that map to previous data we read from
the file. Calling file:advise/4 after every read significantly
slows down the btree build step, therefore we call it after
every 100Kb bytes were read.

This is a simple adaptation of the technique employed in rsync:

    http://insights.oetiker.ch/linux/fadvise/

Verified that page cache size doesn't grow by ~500Mb when building
a btree from a 500Mb file (via /proc/meminfo and vmstat utility).

The file_advise/4 wrapper was added in OTP R14A:

http://www.erlang.org/download/otp_src_R14A.readme (OTP-8637)

Change-Id: Icb96cf5560d83027a0468cd974f3c45f622d05fa
Reviewed-on: http://review.couchbase.org/20623
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>
Reviewed-by: Volker Mische <volker.mische@gmail.com>