MB-6672: Sleep for right amount of time when not on Windows
The UNIX function 'usleep' takes a quantity in microseconds as a
parameter, but we were wrongly passing it a quantity in milliseconds.

Instead of multiplying the period by 1000, to get microseconds, use
intead the UNIX function 'nanosleep', which receives a timespec
structure, as 'usleep' in many implementations doesn't allow values
greater than 1000000us.

Change-Id: I3ee539e2f86d7088f28968650da41bbdb09c2572
Reviewed-on: http://review.couchbase.org/20863
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Volker Mische <volker.mische@gmail.com>