MB-6689 Avoid any work if list of partitions is empty
For calls that request a transition of partitions between the
indexable and unindexable states, totally ignore the call if the
list of target partitions is empty, since it has no effect.
This avoids unncessarily restarting the updater, compactor and
cleanup processes.

It was observed in the logs from MB-6689 that ns_server quite
often passes an empty list of partitions into the functions
couch_set_view:mark_partitions_unindexable/3 and
couch_set_view:mark_partitions_indexable/3.

Change-Id: I59365525e2349fb216711b6cc9fde00ba6e917c8
Reviewed-on: http://review.couchbase.org/21042
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>