MB-6490 Fix race condition in test 20-debug-params.t
After deleting the databases, wait for the view manager
server to process the view group process exit before
attempting to configure/define a new group. This ensures
we don't get the PID of the old view group that was just
shutdown but had its exit message not yet processed by
the view manager.

Change-Id: I95e86a85203f4c4ea9ddc8ca1a3a4717be6dc581
Reviewed-on: http://review.couchbase.org/21266
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>