MB-6804 Ignore effectless indexable/unindexable calls
Because during rebalance ns_server just keeps asking to set a
group of partitions as unindexable/indexable and they are all
already in the requested state. This avoids restarting updater,
compactor and cleaner processes, as well as writing new headers.
Such effectless and useless calls should be ignored and be a
no-op (node side effects).

Change-Id: I7454bb9b6445db98f7d9111e769f5b566cc2831e
Reviewed-on: http://review.couchbase.org/21304
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>