Don't release binary after it's transferred to a term
If compression is successful, the binary containing the
compressed content is transferred to an Erlang term.
Yet, the destructor of the compression sink class will
release the binary - while this doesn't seem to cause
any problem, the OTP man pages don't mention if it's safe
to release a binary that was transferred to an Erlang term.
All the OTP man pages say is that a binary is freed by
either calling enif_release_binary or enif_make_binary,
giving the idea they are mutually exclusive operations.

Change-Id: I341f685973031ae13b71daacd3febaf358c2f3be
Reviewed-on: http://review.couchbase.org/21339
Reviewed-by: Aaron Miller <apage43@ninjawhale.com>
Tested-by: Aaron Miller <apage43@ninjawhale.com>
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>