Don't allocate binary buffer with 0 bytes
For the standard C library, the behaviour of calling malloc
to allocate a bufer with 0 bytes is implementation specific.
Some implementations return a valid pointer while others
return a NULL pointer.

It's not clear if the same problem can happen directly with
enif_alloc_binary(), or if its implementation can result in
an malloc call for 0 bytes.

Nevertheless it's safer to start with an initial binary buffer
size greater than zero.

Change-Id: I282d4fab6d86be478d0cd5233c433e4b25b6e5eb
Reviewed-on: http://review.couchbase.org/21340
Reviewed-by: Damien Katz <damien@couchbase.com>
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>