MB-6805 Notify update listeners of index updater error
It's possible that several classes of errors prevent the
indexer from progressing (like failures opening database
files for example), which means it blocks update listeners
forever, and this prevents rebalance from progressing.
For this scenarios just send an error message to the
update listeners, so that rebalance doesn't get blocked
forever.

Change-Id: I3c97296d4a06c698917b085b713c547649f18563
Reviewed-on: http://review.couchbase.org/21382
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>
Reviewed-by: Damien Katz <damien@couchbase.com>
Reviewed-by: Volker Mische <volker.mische@gmail.com>