MB-6629 Faster index compaction retries
Revert to initial approach where compactor keeps
reading log file until it reaches EOF, instead of
coordinating with the updater regarding where the
EOF is and stop once it reaches that EOF, which
very often no longer matches the real EOF.

Verified with evperf lucky8 test that with this
approach there's no more than 2 to 3 retries, as
opposed to up to ~20 retries with previous approach.
Total compaction time average, and 99% percentile,
is almost half of what it was before during the
lucky8 test run.

Change-Id: I3c893a780a7878080fb366ba72b15e544bb36834
Reviewed-on: http://review.couchbase.org/21516
Reviewed-by: Farshid Ghods <farshid@couchbase.com>
Tested-by: Farshid Ghods <farshid@couchbase.com>