MB-6995 Don't restart couch_{server,set_view} on dir changes.
If some external (meaning ns_server) process calls anything from
couchdb at the time of restart, it may cause this process to terminate
in the hard way with reached_max_restart_intensity error. To avoid
this we're going to stop couchdb from ns_server in coordination with
all the other ns_server processes.

Change-Id: Ifea0c6e3bbebbdd36efedd4fe10bb2a7ccb27892
Reviewed-on: http://review.couchbase.org/21946
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Damien Katz <damien@couchbase.com>
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>