MB-7030 Always send new group snapshot to parent
Instead of sending it with a minimum periodicity of 5 seconds,
every time a batch of changes is applied to the btrees, send the
new view group snapshot to the parent. This prevents discarding
some indexing progress when there's a quick succession of partition
state transitions, as it happens during rebalance.

This restriction was added in the past with the goal of reducing
memory copying (group snapshots were structures with much more
fat).

Change-Id: I4a56b1eaf1019e3203dcaf4e23422145a0b64d21
Reviewed-on: http://review.couchbase.org/22006
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>
Reviewed-by: Damien Katz <damien@couchbase.com>
Tested-by: Damien Katz <damien@couchbase.com>