MB-7030 Don't fsync when doing some header commits
Fsync after writing an header only if the new cleanup bitmask
differs from the old cleanup bitmask. This is safe nowadays
as ns_server on startup configures view groups (MB-6310), and
sets the correct vbucket states. The fsync when cleanup bitmask
changes is necessary because otherwise on restart after a crash,
the view group might attempt to open a vbucket database that was
deleted right before the crash but before the header hit the disk.

Change-Id: I6b70cb7fae55f6a856c56aae8659e8072a64b351
Reviewed-on: http://review.couchbase.org/22010
Reviewed-by: Damien Katz <damien@couchbase.com>
Tested-by: Damien Katz <damien@couchbase.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>