MB-7039 Avoid crash if replica index file is missing
When opening a previously created and configured index, if
for some reason the replica index file is missing (or doesn't
have an header, due to a crash), we reach a situation where
opening the index becomes impossible, unless we explicitly
delete the main index file.

The solution here is to configure the replica index, if it's
missing, when opening the main view group.

Haven't observed this happening in practice, but it's a possibility, therefore good to address for extra reliability.

Change-Id: I42fe0276770a4469a83a94e5e0f88bbbd5370d99
Reviewed-on: http://review.couchbase.org/22038
Reviewed-by: Damien Katz <damien@couchbase.com>
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Damien Katz <damien@couchbase.com>