MB-7082 Revert "MB-6107 Remove no longer needed and wrong logic"
This reverts commit 29ea3906f80dfa7a4487ba7cb2e5308877f635c7.

Turns out that for some periods of time, the same partition can be
marked as active in 2 nodes. Therefore for one of them, the group
snapshot bitmasks have to be manipulated, so that we get results
for that partition from only one node.
Due to lack of full atomicity, ns_server passes a list of wanted
partitions for each node which deals with this case. These lists
of wanted partitions must be strictly respected.

Change-Id: Ie6cc2464ff13b196c4bbefda90693d34d89f0bb2
Reviewed-on: http://review.couchbase.org/22251
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>
Reviewed-by: Volker Mische <volker.mische@gmail.com>