Revert "MB-6107 Remove no longer needed code"
This reverts commit c148470cf60a3d8ecff617cbb71a95091efed11b.

It's still needed. While the compactor is running, the updater
might be transferring the replicas partitions (happens right
after failover) and finish before the compactor finishes. Later,
if the compactor finishes, we accept a snapshot where the replica
partitions data is not signaled as transferred. This will self
heal over time however, but it make take some considerable amount
of time to heal.

Change-Id: Id0dcf0a6dcfd4904e3e9d1283e229f59047639b2
Reviewed-on: http://review.couchbase.org/22312
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>
Reviewed-by: Volker Mische <volker.mische@gmail.com>