MB-7702 Fix groups received from the compactor
Due to the interaction with the updater, the compactor
might receive a group snapshot from the updater that
doesn't reflect the state of the snapshot in the parent
view group process. Therefore, in the view group process,
before accepting a group from the compactor, fix that
group so that it reflects the latest state, just like what
is done when the view group process receives a group
snapshot from the updater.

This could lead to a permanent bad index state (missing
passive partitions, or partitions marked as passive when
they should be active and vice-versa). This was a regression
introduced by MB-7522:

"Optimize very common index state transitions (rebalance)"

Change-Id: I2637d0fda838e803b5bcd474a86fb02c16cfe992
Reviewed-on: http://review.couchbase.org/24534
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>