MB-7813 Allow replicas on transfer to be unindexable
Previosuly it was not possible to move to the unindexable state
partitions that are replicas being transferred into the main
index (marked as passive). This change now allows it, as from
an external point of view, once a replica is requested to become
active, it's seen like a regular active partition - the details
of transferring a replica are meant to be known only internally,
abstracted from callers.

Change-Id: Ife72afe0e6827427a23b4bf48fe2eb6a48481758
Reviewed-on: http://review.couchbase.org/24946
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>