MB-7882 No id btree lookups for dynamically added partitions
If a partition is dynamically added to the updater while
it's running, add it to the list of new partitions, so
that id btree lookups aren't performed for documents
comming from that partition - just like what is done for
new partitions that were added before the updater started.

This only has impact under high memory pressure scenarios,
because the same btree nodes will end up being visited
later by the pipeline stage which updates the btree.

This should have been included as part of the MB-7522,
commit ac82c60302422747e8804d566211a20684ec78fb:

"Optimize index state transitions (rebalance)"

Change-Id: If47f821445d0fa5e98ec74d89ea79d1b0a91c45e
Reviewed-on: http://review.couchbase.org/25050
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>