MB-7978 Fixes to view tree with multiple reducers
A view tree with multiple reducers where not all of them
are native (_count, _sum, _stats) or custom (JavaScript)
was not working anymore since the termless storage format,
and it became undetected for so long due to the fact that
old JavaScript testsuite (Futon) was removed (commit
304cbbd93400abfcb07ca70eecb4e1ccf3f3b485) and no equivalent
etap tests existed.

A view tree with multiple reducers only exists when 2 or
more views in the same design document have exactly the same
map function, on a byte by byte basis, and different reduce
functions (builtin or custom).

Change-Id: Id3218842da83e9e26339e6c59dd7c9b2e8d6a999
Reviewed-on: http://review.couchbase.org/25390
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>