MB-7989 Add missing check for pending transition's contents
If a partition is requested to move into the active or passive
state and it's currently already in that desired state and
marked as unindexable, check if it's already in that state
in the pending transition as well - when it is, don't raise
an error, just accept it as it's a noop.

This fixes commit 01d7e11255f8fb2d984604e2383b5b88c5e0a507 and
adds a specific test for this case.

Change-Id: I9d68556f03c0b1d56520a5023f91debfafd0c786
Reviewed-on: http://review.couchbase.org/25438
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>