MB-8029 Use native file sorter for incremental index updates
The native file sorter, part of couchstore, is a C program to
sort files containing view KV records. Instead of doing this in
Erlang, we now do it with the external file sorter, which is more
efficient and faster.

Sample test:

For an index with 4 views (from evperf lucky8 test) and with
3.2M items per view, adding more 3.2M documents and triggering
an index update, the time it took to update the index and the
final index file size were measured.

Before this change:

Index update time:  10m21.888s
Final file size:    4.0Gb

After this change:

Index update time:  7m15.144s
Final file size:    3.2Gb

Change-Id: I91248d25f5200dac8ffe9a1ce28f6a73a85c0eb3
Reviewed-on: http://review.couchbase.org/25592
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>